Migrate to Unity 6 and NetObjects 2.1.1

+++ Game-Blocket/Assets/Scripts/Management/NetworkVariables.cs
@@ -10,7 +10,7 @@ public class NetworkVariables : MonoBehaviour
 	public static NetworkVariables Singleton { get; private set; }
 	#region Multiplayer
 	public static string ipAddress = UILobby.GetLocalIPAddress();
-	public static int portAddress = 7777;
+	public static ushort portAddress = 7777;
 	#endregion
 
 	private void Awake() => Singleton = this;
